+++ src/client/app/common/views/widgets/instance.vue
@@ -1,0 +1,14 @@
+<template>
+<div class="mkw-instance">
+	<mk-widget-container>
+		<mk-instance/>
+	</mk-widget-container>
+</div>
+</template>
+
+<script lang="ts">
+import define from '../../../common/define-widget';
+export default define({
+	name: 'instance'
+});
+</script>
