+++ src/client/app/desktop/views/deck/deck.column-core.vue
