+++ src/client/app/desktop/views/deck/deck.direct-column.vue
@@ -8,7 +8,7 @@
 
 <script lang="ts">
 import Vue from 'vue';
-import i18n from '../../../../i18n';
+import i18n from '../../../i18n';
 import XColumn from './deck.column.vue';
 import XDirect from './deck.direct.vue';
