+++ src/client/app/desktop/views/deck/deck.direct.vue
