+++ src/client/app/desktop/views/deck/deck.mentions.vue
