+++ src/client/app/desktop/views/deck/deck.tl.vue
