+++ src/client/app/desktop/views/home/user/user.friends.vue
