+++ src/client/app/desktop/views/home/user/user.integrations.integration.vue
