+++ src/client/app/desktop/views/home/user/user.integrations.vue
