--- src/client/app/desktop/views/pages/home-customize.vue
@@ -1,3 +1,0 @@
-<template>
-<mk-home customize/>
-</template>
