+++ src/client/app/desktop/views/widgets/index.ts
@@ -9,6 +9,7 @@ import wPolls from './polls.vue';
 import wPostForm from './post-form.vue';
 import wMessaging from './messaging.vue';
 import wProfile from './profile.vue';
+import wCustomize from './customize.vue';
 
 Vue.component('mkw-notifications', wNotifications);
 Vue.component('mkw-timemachine', wTimemachine);
@@ -19,3 +20,4 @@ Vue.component('mkw-polls', wPolls);
 Vue.component('mkw-post-form', wPostForm);
 Vue.component('mkw-messaging', wMessaging);
 Vue.component('mkw-profile', wProfile);
+Vue.component('mkw-customize', wCustomize);
