+++ src/client/app/store.ts
@@ -10,7 +10,6 @@ const defaultSettings = {
 	home: null,
 	mobileHome: [],
 	deck: null,
-	deckNav: true,
 	keepCw: false,
 	tagTimelines: [],
 	fetchOnScroll: true,
@@ -67,8 +66,7 @@ const defaultDeviceSettings = {
 	deckColumnAlign: 'center',
 	deckColumnWidth: 'normal',
 	mobileNotificationPosition: 'bottom',
-	deckTemporaryColumn: null,
-	deckDefault: false,
+	deckMode: false,
 	useOsDefaultEmojis: false,
 	disableShowingAnimatedImages: false
 };
@@ -82,7 +80,6 @@ export default (os: MiOS) => new Vuex.Store({
 		i: null,
 		indicate: false,
 		uiHeaderHeight: 0,
-		navHook: null,
 		behindNotes: []
 	},
 
@@ -105,10 +102,6 @@ export default (os: MiOS) => new Vuex.Store({
 
 		setUiHeaderHeight(state, height) {
 			state.uiHeaderHeight = height;
-		},
-
-		navHook(state, callback) {
-			state.navHook = callback;
 		},
 
 		pushBehindNote(state, note) {
